backend/requirements.txt
bcrypt==4.3.0
email-validator==2.2.0
fastapi==0.115.11
mangum==0.19.0
mongoengine==0.29.1
passlib==1.7.4
python-dateutil==2.9.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
uvicorn==0.34.0
jinja2==3.1.6
cloudinary==1.44.1